backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
alembic==1.17.0
annotated-doc==0.0.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
asteroid-filterbanks==0.4.0
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.3
cycler==0.12.1
einops==0.8.1
fastapi==0.120.1
filelock==3.19.1
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
gunicorn==23.0.0
h11==0.16.0
httptools==0.7.1
huggingface-hub==0.35.3
idna==3.11
imbalanced-learn==0.14.0
importlib-metadata==8.7.0
jinja2==3.1.6
joblib==1.5.2
julius==0.2.7
kiwisolver==1.4.9
lightning==2.5.5
lightning-utilities==0.15.2
mako==1.3.10
markdown-it-py==4.0.0
markupsafe==2.1.5
matplotlib==3.10.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
networkx==3.5
numpy==2.3.3
omegaconf==2.3.0
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
optuna==4.5.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pip==25.2
primepy==1.3
propcache==0.4.1
protobuf==6.33.0
pyannote-audio==4.0.1
pyannote-core==6.0.1
pyannote-database==6.1.0
pyannote-metrics==4.0.0
pyannote-pipeline==4.0.0
pyannoteai-sdk==0.3.0
pycparser==2.23
pydantic==2.12.3
pydantic-core==2.41.4
pydub==0.25.1
pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytorch-lightning==2.5.5
pytorch-metric-learning==2.9.0
pytz==2025.2
pyyaml==6.0.3
requests==2.32.5
rich==14.2.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
sqlalchemy==2.0.44
starlette==0.49.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.8.0+cu128
torch-audiomentations==0.12.0
torch-pitch-shift==1.2.5
torchaudio==2.8.0+cu128
torchcodec==0.7.0
torchmetrics==1.8.2
torchvision==0.23.0+cu128
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
wheel==0.45.1
yarl==1.22.0
zipp==3.23.0
